add messages to chat
