--- src/modules/chats/chats.hbs
@@ -1,31 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta http-equiv="X-UA-Compatible" content="IE=edge" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <link rel="stylesheet" href="/src/default.css" />
-    <link rel="stylesheet" href="./chats.css" />
-    <title>My Chat_Chats</title>
-  </head>
-  <body>
-    <main>
-      <div id="wrapper">
-        <div id="chatsList">chatsList</div>
-        <div id="chat">
-          <div id="contact">contact</div>
-          <div id="messages">
-            {{#bestfriendChat}}
-            <div class="{{isInbox}}">{{msgText}}</div>
-            {{/bestfriendChat}}
-          </div>
-          <div id="input">input</div>
-        </div>
-        <!-- <div id="setBtn">
-          {{> button value=btns.setBtn href="'../settings/settings.html'"}}
-        </div> -->
-      </div>
-    </main>
-    <script type="module" src="/src/modules/chats/chats.ts"></script>
-  </body>
-</html>
