+++ src/pages/chats/chats.hbs
@@ -1,0 +1,16 @@
+
+      <div id="wrapper">
+        <div id="chatsList">chatsList</div>
+        <div id="chat">
+          <div id="contact">contact</div>
+          <div id="messages">
+            {{#bestfriendChat}}
+            <div isInbox="{{isInbox}}">{{msgText}}</div>
+            {{/bestfriendChat}}
+          </div>
+          <div id="input">input</div>
+        </div>
+        <!-- <div id="setBtn">
+          {{> button value=btns.setBtn href="'../settings/settings.html'"}}
+        </div> -->
+      </div>
