+++ src/pages/chats/chats.html
@@ -10,22 +10,10 @@
   </head>
   <body>
     <main>
-      <div id="wrapper">
-        <div id="chatsList">chatsList</div>
-        <div id="chat">
-          <div id="contact">contact</div>
-          <div id="messages">
-            <!-- {{#bestfriendChat}}
-            <div class="{{isInbox}}">{{msgText}}</div>
-            {{/bestfriendChat}} -->
-          </div>
-          <div id="input">input</div>
-        </div>
-        <!-- <div id="setBtn">
-          {{> button value=btns.setBtn href="'../settings/settings.html'"}}
-        </div> -->
-      </div>
+      <main>
+        <div id="app"></div>
+      </main>
     </main>
-    <script type="module" src="/src/modules/chats/chats.ts"></script>
+    <script type="module" src="chats.ts"></script>
   </body>
 </html>
